Prevent unnecessary alignment code from being used in AthSimulation jobs
The unit tests for the `EventAthenaPool` package pull in various alignment tools
in the configuration even though they are not used. This doesn't matter in Athena
builds, but breaks AthSimulation builds where these tools are not present.
This commit adds some protection to ensure that no attempt is made to load the tools
in AthSimulation builds. Should fix ATLASSIM-3848.


Former-commit-id: f0c213777644e5536bd34e0ddacb0f82f94c1a25