Merging Geant4Tool and G4TransportTool
The `G4TransportTool` is only used by the `Geant4Tool` `ISimulatorTool`
implementation, so it makes sense to merge the two tools into one.

The next steps are to remove the, now obsolete, ISF_Geant4Interfaces
package, then look again at the differences between the merged
`G4TransportTool` and `G4AtlasAlg` to attempt to unify the code.


Former-commit-id: 0434379d495c522cef37ce31b0511c74965db1a7