StoreGate: Remove locking from SGImplSvc::createObj.
SGImplSvc::createObj was taking the store mutex.
This was there originally, but is probably not really needed ---
both DataProxy and the I/O components have their own locks.
Further, this was observed to cause deadlocks for the detector store,
and would in general be expected to be a contention issue.
Remove it and see what happens.

(Deliberately left commented-out rather than removed entirely.)


Former-commit-id: d168f5b940584b3ad0931cbc884a18a8171aa5e5