DataModelRunTests: Fixes for new plugin loader.
The V2 Gaudi plugin loader is fragile in that it can sometimes mysteriously
fail if dynamic libraries have been loaded during the python part
of initialization.  We were doing this to pick up the dictionaries
we needed.  Restructure this as an algorithm in order to defer the
dictionary loading to initialize().



Former-commit-id: 507d09216095afa738511c3427c9ee076b0182c0