Modify Input of MuonRecRTT and Sweep Updates for MuonGeomRTT
This commission is to sweep MR14374 onto branch 21.0 and 21.1.
MR14374 shrink the size of input datasets for MuonRecRTT and create empty pool files
automatically for MuonGeomRTT, and two new tests (test_muonrecrtt_cosmic.sh,
test_muonrecrtt_toroid_off.sh) are added to MuonRecRTT.

In this commission, the test test_muonrecrtt_toroid_off.sh is modified to use
parallel-running mechanism, and the number of input files is changed from 3 to 8.


Former-commit-id: 36913ade6b80bff80ef7e20bbef6ae2586413452