Improve ROS access simulation in athenaHLT (see also ATR-17265 and ATR-18467)
The present simulation was changed to report in addition to the ROS hit rate also
the transfered data volume. Further the simulation writes out, if requested, for
rejected and accepted events a file which contains information for every ROS access.
The data in these files can be used for further analysis.


Former-commit-id: 78e373154536090762744d14d9ea3f6b4fcde817