G4AtlasMTRunManager: initialize geometry before setting smartlessness.
If the detector geometry service is not initialized before retrieving
the logical volume store, the setting of the smartlessness will fail
silently. This commit moves the initialization beforehand. The new order
mirrors the order of the serial version, G4AtlasRunManager. If the
smartlessness is not set, Geant4's default value is used. This results
in a different voxelization, which was the cause of the safety
calculation discrepancy between serial and MT reported in ATLASSIM-3630.


Former-commit-id: 2b88c3f395ec5ade4cc8bec55cac9abc493123b9