ISF InputConverter: retrieve world solid locally
ISF_to_G4Event relies on isInsideG4WorldVolume to retrieve the world
solid and store it as a member variable of InputConverter.
InputConverter is shared amongst threads, so the world solid member
variable is not guaranteed to be thread-safe. The world solid is now
retrieved inside ISF_to_G4Event as a local variable and then passed to
isInsideG4WorldVolume, instead of caching it as a member variable. This
was uncovered in the process of investigating the crashes documented in
ATLASSIM-3861.


Former-commit-id: 1f60f83ab913babcaadcba3973da9d827c4da9ce