Muon truth fixes, mostly NSW related
Neither the MuonTrackTruthTool nor the MuonDetailedTrackTruthMaker were using NSW truth hits.
For the MuonDetailedTrackTruthMaker, an NSW flag was added to determine whether to load the NSW hits or the CSC hits.
For the MuonTrackTruthTool, it proved easier (and more compatible with the new framework) to move the retrieval of objects from SG to the algorithms calling it.
Thus the MuonTrackPerformanceAlg and MuonSegmentTruthAssociationAlg now retrieve the necessary collections, using an NSW flag to determine what hits to load.
The MuonTrackTruthTool and its interfaces are also updated accordingly.
Finallly, the configuration of the algorithms is adjusted to ensure that the NSW flag is set correctly.


Former-commit-id: a45c3847f59fac1fc2473745b112613c649219f6