Switched to using boost::thread_specific_ptr.
While AthContainers_detail::thread_specific_ptr was a better choice in general,
since AthContainers would also like to access TBB when switching it to thread
safe mode, it was easier to just use the Boost type directly in xAODCore.

Note that the assumption is that if we only use ROOT::RDataFrame for multi-threading
in 21.2, not switching on thread safety for AthContainers should be okay. Since all
singleton objects should get set up in memory before the multi-threaded access to
them begins.


Former-commit-id: 27e50454e69d698081e99c3e8089c90aeebe245a