Made xAOD::Internal::THolderCache::getClass(...) have a better interface.
Previously it was not possible to tell whether a returned null pointer meant that the type was not encountered yet, or whether we already tried to find a dictionary for the type, one is just not available. Which could lead to TClass::GetClass(...) getting called by the code over an over again in some cases.

The new interface returns a pair instead, whose first element denotes whether a dictionary was set in the cache already.

Did not document the code too much, since this is not a publicly visible class of xAODRootAccess.


Former-commit-id: 40aad00113b4cff2667e4121c57738fd3db3d958