Adding VeryLooseLLP cleaning and cleaning up
Adding the very loose LLP jet cleaning working point, used in a number
of SUSY LLP analyses (it should be centralized and provided by a tool).
Adding also a tool that applies the cleaning to the extended Jet/MET
setup for derivations.

Taking the opportunity to do a little code cleanup.  No need to keep
things around in comments that don't exist any more.  Fixing up the
README a bit as well.  Also added the doc to the twiki that the readme
points to.


Former-commit-id: c13794572261cd35ed4718bf241d8a542776fc55