Fix dependency of CaloBadChanTool (ATLASRECTS-4635)
TileBadChanTool tools was not declared as property
of CaloBadChanTool and it's dependencies was not propagated
up to CaloBadChanTool. This caused problems in AthenaMT
because this dependency was not propated to the algorithms
which use CaloBadChanTool. Now TileBadChanTool is declared
as property of CaloBadChanTool and made private.
The same is done also for LArBadChanTool because anyway
it should be made private at some moment.


Former-commit-id: 069ece005218e44bd11ed1a661f1a0ca2ccedb36