TileByteStream: Fix size overhead calculation for L2 unpacking.
In 2005, the number of words in the fragment header changed from 2 to 3.
This is indicated in TileROD_Decoder via a member variable.
However, this is not handled correctly in unpackCollectionL2,
which tests the data version but does not set the member variable.
The upshot is that if you're unpacking the new data format,
and you haven't previously unpacked any tile data, then
unpackCollectionL2 will put an extra garbage word in the output.

Fixed with this change.  (Transitional, as we're moving to get rid
of mutable member variables such as m_sizeOverhead.)


Former-commit-id: c9feb823faa5e6bfe20567e36aa240f159f1b096