G4Atlas(MT)RunManager: messages for volume store
Add error message if the logical volume store retrieved from G4 is
empty and info messages if either the Muon::MuonSys and LAr EMB STAC volume
is not present in the store. If the store is missing a volume, the
smartlessness will not be set for that volume. Before, this behaviour
was quiet.


Former-commit-id: a97eb06340ceff722d887738b3fb0520ff47a4fd