Update of DSID ranges attributed to SUSY samples
Hello,

This MR fixes an inconsistency in the DSID ranges allocated to SUSY.
This file was updated:
https://svnweb.cern.ch/trac/atlasoff/browser/Generators/MC15JobOptions/trunk/share/Blocks.list
and as a consequence, we need to update the hardcoded list of DSIDs we use to detect whether this is a SUSY sample.
We may try to find something less hardcoded, but for now we need this to be fixed.

Cheers,
Bertrand


Former-commit-id: 1233f22680149e56c58b5a6491a77affeacc3f2d