Fixed the logic of accessing auxiliary containers directly in kAthenaAccess mode.
The code flow in kAthenaAccess mode is just way too complicated. :-(
Previously, when accessing an auxiliary object directly (so, not through an
interface object), the dynamic store of the auxiliary store did not get
set up correctly. It got deleted by the input TTree, with a pointer in
TEvent still pointing at it.

This change should fix this logic issue. In a not completely trivial manner...


Former-commit-id: 4e64e26c7443c348c35742bccf496f222b871d18