Made it possible to access auxiliary containers in DAOD files through IProxyDict.
In DAODs the type advertised by xAOD::EventFormat for any given branch may not
be correct. The slimming applied during the derivation may have replaced specific
auxiliary container types with xAOD::AuxContainerBase.

To get around this, removed the code'd dependence on the type given by
xAOD::EventFormat, and now it uses the same logic for finding the correct
type as xAOD::TEvent does through its "native" interface.

To make sure that this does not break in the future, added a unit test
for this behaviour. (Which I actually used for developing this fix.)


Former-commit-id: 16589aec561f2f37cea21f6f6703ad8ba5fab5a8