InDetRecExample: Fix tool initialization problem.
A previous change made monitoring tools private, and as a consequence
removed the setting of those tools in ToolSvc.  However, the same
was also done for a few tools which are not (yet?) private.
This was seen only for cosmics configuration, but was causing the
configuration to crash.
Restore the setting to ToolSvc for those tools.
Fixes configuration crash in q220.



Former-commit-id: ceb5bfc84ae0a81e485b9a793ac35a366f777d12