JetTagMonitoring: Fix problem with private tool configuration.
A previous change made monitoring tools private.
But one must be careful:

```
tool2 = Tool2(...)
tool1.Tools += [tool2]
tool2.foo = 1
```

If the tools here are private, then when we append to tool1.Tools,
the tool2 instance is _copied_ to a new, private instance.
Hence, the following setting of foo does not affect the private 
instance used from tool1.  If the tools are public, however, then
the setting of foo does affect the public instance referenced by tool1.

This was breaking the configuration in JetTagMonitoring_jobOptions.py.
Fix by moving the appending of the tool to the end.



Former-commit-id: 4948ca5290fca28ddbecd96213d5a09653a0bba2