Fix a bug in the nominal jet smearing
There was a bug in JetCalibTools where the smearing factor was being
taken as sqrt(A-B) instead of sqrt(A*A-B*B).  This has been fixed.
It was confirmed that this only affects JetCalibTools, as the
JetUncertainties package is subtracting in quadrature as intended.

In addition, a small protection was added to both JetCalibTools
and JetUncertainties, preventing a negative smearing factor.  This
is extremely rare (it is many sigma into the tails), but it should
be handled all the same.  This is now done - the smearing is run
until a positive smearing factor is obtained.

Finally, the default CalibArea locations have been updated for
both JetCalibTools and JetUncertainties to point to the latest
recommended areas.  This has zero impact so far, but it will be
important when people start using the new large-R recommendations.


Former-commit-id: 7432c2ac6f6f2cd26b3bbbd6d49f594a710bcad3