opt-out for skipping CollectionTree-less inputs in AnalysisTop
AnalysisTop currently skips input files without a CollectionTree when
running on real data. This appears to be a workaround for a problem
that may no longer exist. It breaks bookmarking when a file without a
CollectionTree is encountered.

Let's treat this workaround as an experimental feature, so one can turn
it off by adding `SkipInputFilesWithoutCollectionTree` to the
`KillExperimental` configuration option.


Former-commit-id: d0b2585813a3b699f369e3b3fe2159573724b1c6