StoreGate: Avoid binding reference to temporary.
In WriteMetaHandle, we had the initialization

```
    m_sid( MetaContBase::SourceID("") ),
```

where m_sid is a reference.  This doesn't work because what we get from
MetaContBase::SourceID("") is a temporary object that is deleted once the
statement completes.  Thus we're left with a reference to a deleted object.

Change so that m_sid holds a value, not a reference.


Former-commit-id: 5e48654b59164205d1d5fd8203a6d82a782de2ac