First attempt at TruthLinkRepointTool
There are a number of cases where reco objects have a decoration that is
a link back to a truth particle. In some cases, these truth particles
live in large collections that are dropped during derivation making.
This code provides a tool and a default python configuration that will
re-point these links from the small collections over to the mini truth
collections that are included in the derivations.


Former-commit-id: b0e25c8a3401fbc6fd5881d678c74bcdd8ed67a4