Decrease dependency on TileInfo
TileInfo is used to return any detector-dependent
constants or information regarding the Tile Calorimeter,
including conditions data via different conditions tools.
It is not good for AthenaMT.

So functionality of TileInfo which is provided via tools
will be removed and clients will be migrated to use tools directly.
It is supposed that TileInfo will keep and provide only constants
which will be initialized during initialization of job.

Migration of clients has been started.


Former-commit-id: a571f58e6316d72c3e9474b215ce78fbcbf6e99a