HION derivation modifications (ATLASDPD-1183)
The detailed contents for electrons and photons as defined by EGamma CP
were added to HION5. This is necessary to study shower shapes in HI
enviroment and to be able to recompute the electron ID optimized for
HI data at analysis level.

Also, the identification of 2018 Pb+Pb datasets was added in HISkimmingTools.py.


Former-commit-id: cffa9f6089a597329239156c4bfe573d74afdaf0