xAODTracking: Specialize SG::Zero to avoid uninitialized fill() with Eigen.
Eigen objects have the annoying property that the constructor does not
initialize the object.  Internally, DataVector can call fill() on an
aux variable vector, passing a default-constructed instance.
In the case that an Eigen object is used directly as an auxiliary variable,
we can then get warnings about the use of uninitialized data.

Specialize SG::Zero for Eigen::Matrix so that the instance gets
fully initialized.


Former-commit-id: 931765d9024d6168f0b5108bdd2db60d7225e56e