Introduced the xAODMaker::EventInfoNonConstCnvAlg algorithm.
In order to allow running jobs that read an input file holding xAOD::EventInfo,
which themselves further modify this object, introduced the new
xAODMaker::EventInfoNonConstCnvAlg algorithm. Which makes a deep copy of the
original object, and overwrites it in the event store with its deep copy.

To test that the code works as expected, introduced xAODReader::EventInfoModifierAlg,
an algorithm testing that a non-const xAOD::EventInfo object can be retrieved, and then
modified. And added a jobOption testing that the two algorithms can work correctly.


Former-commit-id: dc5c875e9e1a2b4e50e1fb1ba428deb995e69c76