IOVDbSvc: Implement extensible folders.
Allow a folder to be tagged as `extensible'.  In that case, if we read an open-ended
range, then we set the end time to just past the current event.  The end time
will be automatically updated on subsequent events as needed.  This allows
the later addition of another IOV range using IOVSvc::setRange.
This is needed by HLT; see ATEAM-477.


Former-commit-id: f7b73216f1e96364b237ed676757754af697249c