Fix matching of merged NSW PRD to SDO
For the NSW, one PRD is made for one readout channel. A cluster is formed of many readout channels,
so many PRDs "merged PRD". Usually, the truth matching retrieves an identifier from a PRD and checks
whether the channel number corresponds to the SDO. However, this doesn't work for NSW merged PRD.
This commit fixes this problem.


Former-commit-id: 67741986c1d6136ce19003c7fe7c13d79110061c