AthenaConfiguration: Don't crash if downstream packages are missing.
AthenaConfiguration is a package that's needed by many other packages.
As such, it should not be depending on things like LArCellRec --- that
results in a circular package dependency.
Why should running the SCT_Cabling tests require LArCellRec to be present?
It shouldn't, but it currently does.

Work around this by not crashing if downstream packages are not present.


Former-commit-id: ae66736f067fc6cb2fd53b033b941bc24a327007