Adding back AntiKt4HIJets to HION5 output (ATLASDPD-1183)
In older versions of the HI reconstruction, an AntiKt4HIJets container
was stored in xAODs, while a container with corrected jets (DFAntiKt4HIJets)
was constructed on the fly during derivation production.
The current version of the HI reconstruction used in Tier0 during the
2018 Pb+Pb data-taking applies the corrections to jets already during xAOD
production, and there is no need to reconstruct DFAntiKt4HIJets during
derivation production. Therefore, the AntiKt4HIJets need to be written to
the HION5 output explicitly.


Former-commit-id: 9516a9aad89009ba43efccf83323855d98e5edfe