Add possibility to filter Offline tracks by Roi for the ntple creation
Added options to the ntple filling code, such that the Offline tracks can
be filtered so that only those inside an Roi will be retained.

This is disabled by default, and can be enabled with a new job option for
the TrigTestMonToolAC.

It reduces the size of the Rois by afactor of around two.


Former-commit-id: 4ea01c330368f9aa4b33f32bdcecfcffc4de36e5