Remove unnecessary ParticleBroker member variable from FastCaloSimSvc and FastCaloSimSvcPU
The `BaseSimulationSvc` has a member called `m_particleBroker` which is a pointer to `IParticleBroker`.
The `FastCaloSimSvc` and `FastCaloSimSvcPU` derived classes have a `ServiceHandle<IParticleBroker>`
member variable with the same name. The member variable in the derived class serves no purpose as
the base class member vairiable is set directly by a call to `ISimulationSvc::SetParticleBroker` in
`SimKernel`. This is confirmed that running a frozen AFII simulation test on these changes showed
no change in simulation output.


Former-commit-id: c1d8845429dcc2711e137d305a7ab47adcd36a8d