fixed initialization problem when multiple instances of the InDetTrackSmearingTool are created
Several histograms containing the smearing factors are loaded when the tool is invoked for the
first time. The boolean flag keeping track of this was however defined as "static", i.e. when
multiple tools are created, only one is ever initialized.

Implemented solution: change the flag to a private variable.


Former-commit-id: a3b8e94005535e07ea9fda5d82255e96be3ae1b8