Optimize the BPHY15 derivation
Add chi2 cuts in the three cascade algorithm JpsiPlusDsCascade, JpsiPlusDpstCascade and JpsiPlusDs1Cascade. To improve the running speed, the Jpsi, D_(s)+, D0 and K_S0 candidates are pre-selected outside the loop of cascade fit. The charge requirements in the D+/-, D*+/- and D_s1+/- modes are re-written so that both positive and negative modes could be selected in one instance in BPHY15.py. The cuts are also adjusted to reduce the numbers of combinatorail backgrounds.


Former-commit-id: d612713789b031b3ec7e58b4257126e2ad9d1fe9