Add safety for HDF5 output types that are iterable
The static assertions that checked for type mismatches between the
consumers and the arguments to `fill` were a bit aggressive. In
particular, if the output type given to the consumer was also
iterable, the checks would recurse into the iterable until no more
iteration was possible, and conclude that the types and depths didn't
match.

This adds some protection against that situation.


Former-commit-id: 2feb17f5da9d81a20dbdc45e945be2c9a895fe43