Adding tau seed jet width to DFSUSY
Hello,

This MR is a resubmission of !15423 which was targeting DerivationFrameworkTau.
The functionality to decorate taus with their seed jet width is instead added to DFSUSY because we can't wait.
A new common python file is added, in order to schedule the decoration across several SUSY formats.
I used a CommonAugmentation in order to avoid defining a decoration name that would be specific to each format, to avoid interferences when running in a same carriage.
The decoration is added to SUSY2, SUSY3, SUSY5, SUSY11 and SUSY18.

I also removed a number of tau ID variables from SUSY11: the pileup-corrected variables are no longer used as input to the BDT, now that the BDT score is flattened vs mu.

There are also a few new jet and trackjet collections in SUSY11, these were implemented by Walter and merged internally with this MR to gain time.
For SUSY11, the change in output size is +14.5% for data (SUSY11 is almost exclusively for data).

I set the 'urgent' label because this needs to go in the derivation cache that will be built on Friday, so that we can launch urgently the DAOD reprocessing of relevant formats.

Thanks.
Bertrand


Former-commit-id: e39460869fb502fa3a57846d21b94ca8420d2a30