Redifine class methods as private
Redifine class methods that are only used by class itself as private
methods in the headerfile. This is to make it in line with the
interface header 'ISCT_RodEncoder.h', so that the methods that are
defined in the interface header are labeled public while the other
methods are labeled private.


Former-commit-id: 38f1f06c0ef57530e2124f96278e5f9312107b29