Adding extra commands for fundamental particles
This was a subtle one.  The width for a particle is taken from the SLHA
file when it's put in the SLHA file. For fundamental particles, they do
appear in the SLHA file input, and here we want to make sure that we
override the width.  In case the width override doesn't work, we also
explicitly set the min and max mass for generation to just either side
of the actual mass.

When sampling the tails of the mass distributions, this could cause
errors in the subsequent decays, where the gluino mass was sampled to be
higher than the RHadron mass.


Former-commit-id: c0bbf3c3f73f22242ae5640e8d004e92cfb6deeb