AtlasUnixGeneratorJob: Protect adding of McEventSelector
Adding the `McEventSelector` only works in case there wasn't
already an `EventSelector` of different type configured before.
Otherwise the `ConfigurableDb` will complain that one tries to add a
Configurable with the same name but different type.

Added a protection and warning message for this case. This will allow
the use of some job options in the HLT (e.g. HelloWorld) where a different
EventSelector is used. For all other cases, this change should have zero
impact as the code was already non-functional before.


Former-commit-id: d53096a49a12c2307c98d1cdc7236cc6b1441201