Updated to VgammaOR tool so that it no longer needs to maintain state.
Instead of setting inputs with a separate function call, all neccessary inputs are now given as argument to the overlap functions.
As a consequence, these functions now need to return a status code and the result is communicated via reference argument instead.
The number of public functions in the header is further reduced, only three different option to calculate OR info and a Frixione
isolation function remans.


Former-commit-id: 536f20c340af83f7e2419e2ae7ef94e6cdb1fe53