Remove usage of deprecated AthenaServices.SetFatalHandler
Do not use `AthenaServices.SetFatalHandler` but instead set the
equivalent property on the CoreDumpSvc. Switch the remaining job options and
delete any comments referencing the old class.

A nice side effect is that the initial bootstrap of `athena.py` is now 3
seconds faster due to the removal of the `cppyy` import (of course this
time is spent later once the import happens at another place).


Former-commit-id: 021e1688626ce68e5faa67683820f68c789f5844