Adding EF combined muons to trigger test
Added the needed algorithms to create EF combined muons for the trigger, and added EF combined step for the muon menu test (the mu.with.views test will be updated in a separate MR later).
Also added some protection in the combined muon hypo for the case where we don't find any combined muons.


Former-commit-id: 98c8d0db5b2e257e6ea3440dddb9bd02f08f1165