Updated AnalysisBase to only build the core/edm packages for now.
To lay the groundwork for future analysis release developments, updated
the AnalysisBase project to only build the most essential packages for now.
So that we could do some further development on top of this core later on.

Removed the RootCore specific code from the project by now. It should
really not survive into Run-3...