DFSUSY: drop tau seed jet width and cleanup
Hello,

This MR drops what had been implemented in https://gitlab.cern.ch/atlas/athenaprivate1/merge_requests/16344, namely the decoration of taus with the seed jet width, because this functionality is being migrated to DerivationFrameworkTau by @lfiorini in !20048 .

I also made some cleanup of old comments and the 'AppendToDict' dictionary, because most collections are now declared in
https://gitlab.cern.ch/atlas/athena/blob/21.2/PhysicsAnalysis/DerivationFramework/DerivationFrameworkCore/python/ContainersOnTheFly.py
and no longer need to be declared in each format.

Thanks.
Bertrand