Decouple BS conversion in TGCs from RDO->PRD decoding tool
In preparation for the migration of the Athena code to the multi-threaded AthenaMT, remove the (scheduler-unsafe) dependency of the TGC RDO->PRD tool on the existence of the RDOs which are in some cases produced from the BS using a decoding function called by the RDO-PRD process itself.

The RawDataProviderTool which can perform the BS decoding is removed from the Muon Spectrometer TGC RDO->PRD convenience tool and can now be used directly by trigger and offline in a similar manner. Changes have been made to the tigger validation unit tests to reflect these changes.