Add track selection to DL2
We train IPRNN with a track selection similar to IP3D. In DL2 there
was previously no way to apply a track selection, so I've added it
with this commit. For now there are two hardcoded track selections:

 - ALL: select all the tracks.
 - IP3D_2018: what we were using in IP3D around the end of 2018.

As with many of the DL2 input behaviors, the track order is currently
determined via regexes that parse the NN file. The list of regular
expressions is currently hardcoded in DL2HighLevel.cxx.