Add Athena Algorithm for interpolation of irradiated E fields
The irradiated E fields are provided from TCAD simulations for a limited
number of values for the amount of irradiation and bias voltage applied
to the sensor. The EfieldInterpolator interpolates these maps to obatain
E fields for any fluence or voltage value of interest. Different
techniques are realized, cubic splines are used as default method. The
input is taken from text files and read into TTrees.