Prevent MS track particles from being added to MuGirl muons
In cases where MuGirl muons are also reconstructed by STACO, they can have an MS track particle, the one used by STACO, associated to them.
Given that this MS track was not used in reconstructing the muon, this can be misleading: this fix ensures that MuGirl muons never have an associated MS track.
Since the MCP tools use the ME track, this change should be transparent for the vast majority of analyzers.
The MS-only extrapolated track, only used by experts for debugging, is retained.