TrigGenericMonitoringTool: copy defineHistogram from AthenaMonitoring
TrigGenericMonitoringTool is deprecated and is only kept in master as
long as we need to keep running the Run-2 trigger. As such it is better
that it contains its own copy of `defineHistogram`. That way we can
freely modify the new version in AthenaMonitoring and don't have to
worry about backwards compatibility.