Updated the worker classes to look for memory leaks in the jobs.
The workers simply check the total amount of memory used by tha analysis process
after initialisation and after finalisation, and then make a difference between
the two. This is a very simplistic approach, which underestimates the memory
leak of the code, but on first order this should be good enough.