Fix some doxygen complient headers, fix the protection of the base class for the particleType class
Some of the head comments were skipped when making fixing the doxygen
compliance, this change fixes those.

In addition, it also fixes the protection of the base class of the
particleType class. This class inherits from ad std::map<int,string>
and ordinarily one should *never* inherit from and std class in this
way.

However, this change makes the base class private so that the map
can never be exposed either from this class or any subsequently
derived class, and so avoids all the very sound reasons why
inheritance like this would otherwise be a very bad idea.