Fix compilation of TrigJspiMonitoring
This TrigJpsiMonitoring is legacy code that is no longer used and should
probably be removed from the release. It uses the base classes from
TrigInDetAnalysis and so this change is to take account of the fix
in the TrigInDetAnalysis headers made for compilance with the coding
rules.

There are still remaining many warnings about deprecated variables such
as the endOfRun, but since all this code will be changed in the future
and I don't see much point in fixing this at this stage, and instead we
should just archive this code, but accept this fix just for the time
being so that the release will at least build.