Add EF combined muons to trigger tests (ATR-17553)
This adds the combined muons to the mu.with.views tets, and also fixes a bug with the input RoIs in that test.
Also updates both muon tests to run over different events so that the combined muon reco actually runs in at least one event (and correspondingly updates the reference for the menu test).