Improved the extra CMake argument handling in the build scripts.
This is to allow us passing multiple extra arguments to CMake through the build
scripts. The formalism is that the -x argument of all build scripts can be
specified multiple times. Resulting in each value passed to -x being taken
into account.

At the same time removed the git version checking from the build scripts,
as it was not needed for anything for a while now.

Finally, took the opportunity to do some general cleanup as well.