create AsgTesting package from AsgTools package
The idea here is that I'd like to avoid linking the gtest-library in
every package depending on AsgTools (which are a lot of packages).
This also required moving a lot of the remaining unit tests from
AsgTools to AsgExampleTools, which already contained a number of tests
for the AsgTools package anyways.