Optimization of CPU time of FastCaloSimV2 (ATLASSIM-3980)
The CaloCellContainer is set to be incomplete by default. Thus, a look up table is needed to to find the correct
index based on the hash value of the cell element. In fact, the CaloCellContainer is always ordered and complete
after creation the look up table is not needed.

After modification, the CPU time/evt is reduced from 71ms to 53ms for single photons.