FastCaloSim: Making FastShowerCellBuilderTool const.
TH1::GetRandom() and TH2::GetRandom2() are not MT-safe, since they're
non-const and also use gRandom.  These are used from
ParticleEnergyParametrization.  Rewrite so as to be const and to allow
passing in a random generator rather than having to use gRandom.
(We currently still pass in gRandom.  Changing that would change rng seeding,
so that will be fixed in a subsequent change.)